main.py
import csv, json
import os
import zipfile
import re

import time, datetime
from xml.dom import minidom

import io, shutil, posixpath
import logging

import concurrent.futures
import copy
import uuid
import traceback

import xml.etree.ElementTree as ET
from bs4 import BeautifulSoup, NavigableString

from PIL import Image, ImageDraw, ImageFont

from tqdm import tqdm
import colorama

