requirements.txt
beautifulsoup4==4.14.3
colorama==0.4.6
pillow==12.1.0
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
